Fixed multiple small errors

+++ dashboard/searchOrder.php
@@ -54,18 +54,20 @@ if(isset($_GET["id"]) && isset($_GET["type"])) {
                             if(isset($_GET["searchBox"]) && !empty($_GET["searchBox"])){
                                 //set the id of the order
                                 $order->order_id =$_GET["searchBox"] ;
-                                //get the order detail
-                                $order->getOrder();
-                                echo "<tr>";
-                                echo "<td>$order->order_id</td>";
-                                echo "<td>$order->order_time</td>";
-                                //display the order detail
-                                echo "<td><a href='orderDetail.php?id=$order->order_id'><i class=\"fas fa - trash text - danger\"></i>View</a></td>";
-                                //delete the order button
-                                echo "<td>
+                                if ($order->exists()) {
+                                    //get the order detail
+                                    $order->getOrder();
+                                    echo "<tr>";
+                                    echo "<td>$order->order_id</td>";
+                                    echo "<td>$order->order_time</td>";
+                                    //display the order detail
+                                    echo "<td><a href='orderDetail.php?id=$order->order_id'><i class=\"fas fa - trash text - danger\"></i>View</a></td>";
+                                    //delete the order button
+                                    echo "<td>
                                                 <a href='displayAllOrders.php?id=$order->order_id&type=delete'><i class=\"fas fa - trash text - danger\"></i>Delete</a>
                                          </td>";
-                                echo "</tr>";
+                                    echo "</tr>";
+                                }
                             }
                             ?>
                             </tbody>
